Velocities are doubles

+++ src/com/MRS/NeckbeardEngine/Enemies/Thug.java
@@ -4,7 +4,7 @@ import com.MRS.NeckbeardEngine.*;
 
 public class Thug extends Enemy {
   
-   public Thug (State state, int x, int y, int xVelocity, int yVelocity, Projectile projectileType, PowerUpPickup powerUpPickup, long timeLine ) {
+   public Thug (State state, int x, int y, double xVelocity, double yVelocity, Projectile projectileType, PowerUpPickup powerUpPickup, long timeLine ) {
         super(state, x, y, xVelocity, yVelocity, projectileType, powerUpPickup, timeLine);
   }
